requirements.txt
asgiref==3.5.0
Django==4.0.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
